Move flashblocks tests to integration suite
