+++ crates/flashblocks-rpc/src/lib.rs
@@ -3,5 +3,3 @@ mod pending_blocks;
 pub mod rpc;
 pub mod state;
 pub mod subscription;
-#[cfg(test)]
-mod tests;
