+++ crates/flashblocks-rpc/tests/assets/genesis.json
